pyproject.toml
[build-system]
requires = ["setuptools"]

[project]
name = "typechange_bot"
version = "0.0.1"
requires-python = ">=3.10"
description = "Type Change Tracking Bot"
readme = "README.md"
dependencies = [
"Django~=4.2",
"celery[redis]~=5.4",
"django-celery-results~=2.5",
"PyGithub~=2.4.0",
"python-decouple==3.1",
"django-treebeard>=4.7.1",
"gitpython~=3.1.43",
"whatthepatch~=1.0.7",
"dj-database-url~=2.3.0",
"pandas~=2.2.3"
]

[project.optional-dependencies]
mysql = [ "mysqlclient~=2.2.7" ]
dev = [
"django-extensions~=3.2",
"mypy~=1.11.2"
]
gumtree_support = [
"parso~=0.8.4",
"tree_sitter==0.21.3",
"pyyaml"
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include=["*"]
exclude=["flow", "notes", "pages", "templates"]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
        'test-data/',
        'survey/migrations/'
]